+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added search function to user setting in myems-admin and myems-api
 - added search function to energy storage container setting in myems-admin and myems-api
 - added search function to energy storage power station setting in myems-admin and myems-api
+- added search function to command setting in myems-admin and myems-api
 - added equipment comparison report
 - added search function to contact setting in myems-admin and myems-api
 - added asterisks to all required fields in myems-admin