+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added search function to shopfloor setting in myems-admin and myems-api
 - added search function to user setting in myems-admin and myems-api
 - added search function to energy storage container setting in myems-admin and myems-api
+- added search function to energy storage power station setting in myems-admin and myems-api
+- added search function to command setting in myems-admin and myems-api
 - added equipment comparison report
 - added search function to contact setting in myems-admin and myems-api
 - added asterisks to all required fields in myems-admin
@@ -17,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - fixed translations of menu setting page in myems-admin
 - fixed value range issues of point in myems-admin
+- fixed pop error issues when binding working calendar to store, shopfloor, and teant in myems-admin
 ### Removed
 
 ## [v5.9.0] - 2025-09-28