+++ CHANGELOG.md
@@ -10,12 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added search function to energy storage container setting in myems-admin and myems-api
 - added search function to energy storage power station setting in myems-admin and myems-api
 - added search function to command setting in myems-admin and myems-api
+- added search function to microgrid setting in myems-admin and myems-api
 - added equipment comparison report
 - added search function to contact setting in myems-admin and myems-api
 - added asterisks to all required fields in myems-admin
 ### Changed
 - updated myems-admin to refresh the point list when binding point to sensor
 - display the username when hovering over the user avatar in myems-web
+- updated myems-admin to refresh the tariff list when binding tariff to cost center
 ### Fixed
 - fixed translations of menu setting page in myems-admin
 - fixed value range issues of point in myems-admin